pyproject.toml
[tool.pdm]

[project]
name = "ifunnygifmaker"
version = "0.1.2"
description = ""
authors = [
    {name = "jaypyles", email = "jpylesbusiness@gmail.com"},
]
dependencies = [
    "python-dotenv>=1.0.0",
    "requests>=2.28.2",
    "pillow>=9.5.0",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-pep517>=1.0"]
build-backend = "pdm.pep517.api"
